% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peranavolley.R
\docType{package}
\name{peranavolley}
\alias{peranavolley}
\alias{peranavolley-package}
\title{\pkg{peranavolley}}
\description{
Functions for reading and working with Perana Sports volleyball scouting files.
}
\author{
Ben Raymond \email{ben@untan.gl}
}
