% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{plays}
\alias{plays}
\title{Extract the plays component from a peranavolley object}
\usage{
plays(x)
}
\arguments{
\item{x}{peranavolley: a peranavolley object as returned by \code{pv_read}}
}
\value{
The plays component of x (a data.frame)
}
\description{
Extract the plays component from a peranavolley object
}
\examples{
\dontrun{
  x <- pv_read(pv_example_file())
  inspect(plays(x))
}
}
\seealso{
\code{\link{pv_read}}
}
