% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{print.summary.peranavolley}
\alias{print.summary.peranavolley}
\title{Print method for summary.peranavolley}
\usage{
\method{print}{summary.peranavolley}(x, ...)
}
\arguments{
\item{x}{summary.peranavolley: a summary.peranavolley object as returned by \code{summary.peranavolley}}

\item{...}{: additional arguments (currently these have no effect)}
}
\description{
Print method for summary.peranavolley
}
\seealso{
\code{\link{summary.peranavolley}}
}
