% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psvt.R
\name{pt_example_file}
\alias{pt_example_file}
\title{Example psvt data files provided as part of the peranavolley package}
\usage{
pt_example_file(choice = 1)
}
\arguments{
\item{choice}{numeric: which data file to return?
\itemize{
  \item{1 - trivial volleyball example file}
}}
}
\value{
path to the file
}
\description{
Example psvt data files provided as part of the peranavolley package
}
\examples{
myfile <- pt_example_file()
x <- pt_read(myfile)

}
\seealso{
\code{\link{pt_read}}
}
