% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psvt.R
\name{pt_read}
\alias{pt_read}
\title{Read a Perana Sports tagger (psvt) data file}
\usage{
pt_read(filename, raw_only = FALSE)
}
\arguments{
\item{filename}{string: path to file}

\item{raw_only}{logical: if \code{TRUE}, just decompress the file, don't parse it}
}
\value{
A list with elements \code{meta} (metadata including the project and template information, and information about the skills and grades used in the tags) and \code{tags} the tag data.
}
\description{
Read a Perana Sports tagger (psvt) data file
}
\references{
\url{http://peranasports.com/}
}
