% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{pv_default_errortypes}
\alias{pv_default_errortypes}
\title{Define errortype interpretations}
\usage{
pv_default_errortypes()
}
\value{
a tibble with columns "skill", "errortype", and "evaluation"
}
\description{
Define the interpretation of errortypes associated with events.
}
\examples{
pv_default_errortypes()

}
\seealso{
\code{\link{pv_read}}
}
