% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{pv_default_subevents}
\alias{pv_default_subevents}
\title{Define subevent interpretations}
\usage{
pv_default_subevents()
}
\value{
a tibble with columns "skill", "subevent", and "evaluation"
}
\description{
Define the interpretation of subevents associated with events.
}
\examples{
pv_default_subevents()

}
\seealso{
\code{\link{pv_read}}
}
