% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{pv_example_file}
\alias{pv_example_file}
\title{Example data files provided as part of the peranavolley package}
\usage{
pv_example_file(choice = 1)
}
\arguments{
\item{choice}{numeric: which data file to return?
\itemize{
  \item{1 - Men's Australian Volleyball League 2017: Canberra Heat vs UTSSU}
  \item{2 - Women's Australian Volleyball League 2017: Canberra Heat vs UTSSU}
}}
}
\value{
path to the file
}
\description{
These example files were kindly provided by Chau Le from Perana Sports.
}
\examples{
myfile <- pv_example_file()
x <- pv_read(myfile)
summary(x)

}
\seealso{
\code{\link{pv_read}}
}
