% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{pv_read}
\alias{pv_read}
\title{Read Perana Sports volleyball data file}
\usage{
pv_read(filename, insert_technical_timeouts = FALSE, do_warn = FALSE,
  extra_validation = 2, raw_only = FALSE,
  eventgrades = pv_default_eventgrades(),
  errortypes = pv_default_errortypes(),
  subevents = pv_default_subevents(), setting_zones)
}
\arguments{
\item{filename}{string: path to file}

\item{insert_technical_timeouts}{logical: if \code{TRUE}, insert technical timeouts at 8 and 16 points}

\item{do_warn}{logical: should we issue warnings about the contents of the file as we read it?}

\item{extra_validation}{numeric: should we run some extra validation checks on the file? 0=no extra validation, 1=check only for major errors, 2=somewhat more extensive, 3=the most extra checking}

\item{raw_only}{logical: if \code{TRUE}, just decompress the file, don't parse it}

\item{eventgrades}{tibble: a tibble that defines the interpretations of \code{eventgrade} values; see \code{\link{pv_default_eventgrades}}}

\item{errortypes}{tibble: a tibble that defines the interpretations of \code{errortype} values; see \code{\link{pv_default_errortypes}}}

\item{subevents}{tibble: a tibble that defines the interpretations of \code{subevent} values; see \code{\link{pv_default_subevents}}}

\item{setting_zones}{named character: if the data file has been scouted using setting zones, then each attack will have its associated setting zone (numbered 1 to 5). The setting zone names are not stored in the file, so they can be provided here as a character vector. This can either be an un-named character vector, in which case it must be of length 5; otherwise if only a subset of the five setting zones are being used then it can be provided as a named character vector(e.g. \code{setting_zones = c("1" = "X1", "3" = "X7", "4" = "medium/fast", "5" = "high")}). The values in this vector will be used to populate the \code{attack_code} column of the returned plays data}
}
\value{
A named list with several elements. \code{raw} contains the extracted but unparsed text from the psvb file, \code{meta} provides match metadata, \code{plays} the play-by-play data in the form of a data.frame, and \code{messages} is a data.frame describing any inconsistencies found in the file.
}
\description{
Read Perana Sports volleyball data file
}
\examples{
filename <- pv_example_file()
x <- pv_read(filename)

x <- pv_read(filename, setting_zones = c("X1", "X2", "X7", "medium/fast", "high"))

}
\references{
\url{http://peranasports.com/}
}
\seealso{
\code{\link{pv_validate}}, \code{\link{pv_default_eventgrades}}, \code{\link{pv_default_errortypes}}
}
