% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pv_recode_attacks}
\alias{pv_recode_attacks}
\title{Generate more meaningful attack_code values}
\usage{
pv_recode_attacks(x, apply_to_skills = "Attack", remap)
}
\arguments{
\item{x}{data.frame or tibble: a peranavolley object as returned by \code{pv_read}, or the plays component thereof}

\item{apply_to_skills}{character: rules will only be applied to rows corresponding to these skills. This function is nominally intended for recoding \code{attack_code} values (hence \code{apply_to_skills = "Attacks"} is the default). But it might be useful for other purposes, in which case other skills might need to be specified here}

\item{remap}{data.frame or tibble: this data.frame must contain the column \code{new_attack_code}, and one or more columns that define the conditions to match on. See the example below for more detail}
}
\value{
A character vector with new \code{attack_code} values
}
\description{
Notes:
\itemize{
  \item "setting zones" (terminology from the VBStats app) are called "attack codes" here for consistency with the datavolley package
  \item the rules are applied in order, so that an attack in \code{x} that matches multiple rows in \code{remap} will have the first matching row applied to it
}
}
\examples{

## read file
x <- pv_read(pv_example_file())

## construct the remapping
remap <- data.frame(attack_code = c("1", NA_character_, "4", "5"),
                    start_zone = c(NA, 3, 2, 2),
                    new_attack_code = c("X1", "Z3", "X6", "V6"),
                    stringsAsFactors = FALSE)
## meaning that: any attack with attack_code 1 will be recoded as an "X1" attack
##               any attack from start_zone 3 will be recoded as an "Z3" attack
##               any attack from start_zone 2 with attack_code 4 will be recoded as an "X6" attack
##               any attack from start_zone 2 with attack_code 5 will be recoded as an "V6" attack

x$plays$attack_code <- pv_recode_attacks(x, remap = remap)

}
