% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pv_tas_data_augment}
\alias{pv_tas_data_augment}
\title{Augment partially-scouted data}
\usage{
pv_tas_data_augment(x)
}
\arguments{
\item{x}{peranavolley: a peranavolley object as returned by \code{\link{pv_read}}}
}
\value{
A peranavolley object with additional rows added to its \code{plays} component.
}
\description{
Augment partially-scouted data
}
\seealso{
\code{\link{pv_read}}, \code{\link{pv_tas_recode}}, \code{\link{pv_tas_live_recode}}
}
