% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{pv_validate}
\alias{pv_validate}
\title{Validate parsed Perana Volleyball scouting data}
\usage{
pv_validate(x, validation_level = 2)
}
\arguments{
\item{x}{peranavolley: peranavolley object as returned by \code{pv_read}}

\item{validation_level}{numeric: how strictly to check? If 0, perform no checking; if 1, only identify major errors; if 2, also return any issues that are likely to lead to misinterpretation of data; if 3, return all issues (including minor issues)}
}
\value{
data.frame with columns message (the validation message), file_line_number (the corresponding line number in the file), video_time, and file_line (the actual line from the file).
}
\description{
Validate parsed Perana Volleyball scouting data
}
\examples{
filename <- pv_example_file()
x <- pv_read(filename)
pv_validate(x)

}
\seealso{
\code{\link{pv_read}}
}
