% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{pv_write}
\alias{pv_write}
\title{Write a Perana Sports volleyball data file}
\usage{
pv_write(x, filename)
}
\arguments{
\item{x}{character: data to write. See e.g the \code{raw} component of the object returned by \code{\link{pv_read}}}

\item{filename}{string: path to file}
}
\description{
This is somewhat experimental. It may be useful if one wants to read an existing file, modify the content, and re-write it back to a PSVB file.
}
\examples{

\dontrun{
 x <- pv_read(pv_example_file())
 new_file_name <- tempfile(fileext = ".psvb")
 pv_write(x$raw, filename = new_file_name)
}

}
\seealso{
\code{\link{pv_read}}
}
