% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{summary.peranavolley}
\alias{summary.peranavolley}
\title{A simple summary of a volleyball match}
\usage{
\method{summary}{peranavolley}(object, ...)
}
\arguments{
\item{object}{peranavolley: peranavolley object as returned by \code{pv_read}}

\item{...}{: additional arguments (currently these have no effect)}
}
\value{
list of summary items
}
\description{
A simple summary of a volleyball match
}
\examples{
x <- pv_read(pv_example_file())
summary(x)

}
\seealso{
\code{\link{pv_read}}
}
