% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pv_tas_recode}
\alias{pv_tas_recode}
\alias{pv_tas_remap}
\title{An opinionated recoding}
\usage{
pv_tas_recode(x, remap = pv_tas_remap(), log_changes = FALSE)

pv_tas_remap()
}
\arguments{
\item{x}{data.frame or tibble: a peranavolley object as returned by \code{pv_read}, or the plays component thereof}

\item{remap}{list: a list with components "conditions" and "values" that define the remapping. See \code{\link{pv_tas_remap}} for an example}

\item{log_changes}{logical: if \code{TRUE}, the returned object will have a "changes" attribute describing the changes that were made}
}
\value{
A modified version of \code{x}.
}
\description{
Note that this function is only likely to be useful if you have scouted your VBStats files following the conventions described at https://raymondben.github.io/scouting-notes/
}
\details{
This function modifies the plays component of a peranavolley object, following the scouting conventions described at https://raymondben.github.io/scouting-notes/. The following changes are made:
\itemize{
  \item attacks marked as "Setter tip" ("Dump" in VBStats) by a setter are mapped to attack_code "PP"
  \item any "Setter tip" by a non-setter are treated as freeballs. Their skill entry is changed to "Freeball", and attack code to NA
  \item attack_code 1 to "X1"
  \item attack_code 2 to "X2"
  \item attack_code 3 to "X7"
  \item attack_code 4 to "CF" (slide)
  \item attack code 5 treated as highball, with new attack code determined by start_zone (V5, V6, etc)
  \item any other attack (with NA attack code) treated as standard tempo with new attack code determined by start_zone (X5, X6, etc)
  \item any dig following a freeball over is changed to skill "Freeball"
}
}
\examples{

## Note: this is a silly example, because the example VBStats file was NOT
##       scouted with the conventions expected by this function

x <- pv_read(pv_example_file())
x <- pv_tas_recode(x, remap = pv_tas_remap(), log_changes = TRUE)

}
