% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{inspect}
\alias{teams}
\alias{home_team}
\alias{visiting_team}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{datavolley}{\code{\link[datavolley:teams]{home_team}}, \code{\link[datavolley]{inspect}}, \code{\link[datavolley]{teams}}, \code{\link[datavolley:teams]{visiting_team}}}

  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}
}}

